+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Properly handle not found (404) errors
 
 ## 0.0.4 (2018-09-04)